// packages/react/src/lib/components/step-indicator/step-indicator.d.ts
import * as React from 'react';
export declare const PdsStepIndicator: React.FunctionComponent<Partial<{
    variant: "horizontal" | "vertical";
    inverted: boolean;
    interactive: boolean;
}> & {
    ref?: React.Ref<any> | undefined;
} & {
    style?: React.CSSProperties | undefined;
} & Partial<{}> & {
    id?: string | undefined;
    slot?: string | undefined;
    className?: string | undefined;
    children?: React.ReactNode;
}>;